chore: organize folder structure
[chore #164523219]
