+++ test/integration/routes/users_test.js
@@ -1,13 +1,13 @@
 import { expect } from 'chai';
 import request from 'supertest';
-import dbHandler from '../../src/dbHandler';
+import dbHandler from '../../../src/database/dbHandler';
 
 let server;
 let user;
 
 describe('USER API ENDPOINTS', () => {
   beforeEach(() => {
-    server = require('../../src/index');
+    server = require('../../../src/index');
     user = {
       email: 'ewere@gmail.com',
       firstName: 'admin',
