+++ test/unit/database/dbHandler_test.js
@@ -2,7 +2,7 @@
 import { expect } from 'chai';
 import config from 'config';
 import jwt from 'jsonwebtoken';
-import dbHandler from '../src/dbHandler';
+import dbHandler from '../../../src/database/dbHandler';
 
 
 let register;
